2nd hand vehicle pricing applicaton
The longest program I ever wrote and its a program which calculates the price of your vehicle using factors like km driven, year of purchase, and the condition of the vehicle. It is having many brands all motorbikes and cars in it with a specific value of the original vehicle. It's gonna calculate what should be the second-hand value of your vehicle.